refactored sevreral files to use the new config
